Refactored how error messages are handled somewhat
fix flow error
